Revise HTML, Add operate() and JS vars
In this commit, I revised the identifiers I've used in the HTML and
divided the screen div into two screens, so that I can eventually add a
portion to the screen which details everything the user has entered, not
just what they are entering in the moment (which the main screen will
contain). From here, I have added a few JS variables, including that
which will determine the contents of the main screen, and I have
declared the operate() function, based on a tested function I developed
in a previous JS exercise. I will test further down if this will work
using the special characters I have decided to use for each operator, or
if I need to revert to using basic operators for the buttons.