Add functions for backspace, decimal
In this commit, I have added similar functions for the backspace and
decimal buttons, both of which are followed up by click-based event
listeners which declare the associated function. The backspace button
should remove the most recently added number from the textContent of
numOne or numTwo when pressed. The decimal button should add a decimal
to the textContent of numOne or numTwo when pressed (though this
function must operate within the parameters of the previous declaration
within the function for the number buttons wherein if numOne or numTwo
becomes longer than 12 digits/characters, the number is too big for the
calculator to handle).