Begin styling
In this commit, I've modified all three code files in order to begin
styling the webpage. A few lines have been added to the JS to change the
font size when the main screen displays an error message (such as when
the number is too long, or when the user attempts to divide by 0). In
the HTML head, I've linked the two fonts I will be using throughout the
webpage (one for the screen and one for the buttons and the rest of the
page). I have also started to add CSS properties to the webpage to start
styling the container, the screen, and the overall page.