Fix kybd divide compatibility
In this commit, I fixed an error I recognized wherein the / key was not
triggering the divide symbol on the calculator as anticipated. A change
to the strictness of equality in that line of JS has resolved this
error. It should be noted that some browsers (i.e. Firefox) also use the
/ key to trigger a "Quick Find" or "Find on page" tool, which is a
slight issue, but I've not yet been able to find anything that can
disable this feature without creating errors in other browsers like
Chrome. Keyboard users can use the "Tab" key to move the keyboard's
input focus back to the calculator.