Button states, backspace character, fix errors
In this commit, I've made edits to all three files to finalize the
webpage for the time being. In the HTML file, I changed the character in
the backspace button to a backspace symbol (⌫) so as to clarify the
button's function. In the JS file, I've added a couple of lines to
correct an error with the backspace button. In the CSS, I have added
hover and active states to the buttons so as to animate them and thereby
make the webpage feel a little bit more interactive and responsive. I
learned the active state and box-shadow properties from a doc on button
animations from W3Schools, which I will link in the README file to cite
my source.